melange index

Library melange

The entry point of this library is the module: Stdlib.

Library melange.belt

This library exposes the following toplevel modules:

Belt A stdlib shipped with Melange

Belt_Array Belt.Array Utililites for Array functions

Belt_Float Belt.Float Utililites for Float

Belt_HashMap A mutable Hash map which allows customized Hash behavior.

Belt_HashMapInt

Belt_HashMapString

Belt_HashSet A mutable Hash set which allows customized Hash behavior.

Belt_HashSetInt This module is Belt.HashSet specialized with key type to be a primitive type.

Belt_HashSetString This module is Belt.HashSet specialized with key type to be a primitive type.

Belt_Id Belt.Id

Belt_Int Belt.Int Utililites for Int

Belt_List Belt.List

Belt_Map A immutable sorted map module which allows customize compare behavior.

Belt_MapDict

Belt_MapInt

Belt_MapString

Belt_MutableMap

Belt_MutableMapInt

Belt_MutableMapString

Belt_MutableQueue First-in first-out queues.

Belt_MutableSet A mutable sorted set module which allows customize compare behavior.

Belt_MutableSetInt This module is Belt.MutableSet specialized with key type to be a primitive type.

Belt_MutableSetString This module is Belt.MutableSet specialized with key type to be a primitive type.

Belt_MutableStack First in last out stack.

Belt_Option Belt.Option

Belt_Range A small module to provide a inclusive range operations [start, finsish], it use a for-loop internally instead of creating an array

Belt_Result Belt.Result

Belt_Set A immutable sorted set module which allows customize compare behavior.

Belt_SetDict

Belt_SetInt This module is Belt.Set specialized with value type to be a primitive type. It is more efficient in general, the API is the same with Belt_Set except its value type is fixed, and identity is not needed(using the built-in one)

Belt_SetString This module is Belt.Set specialized with value type to be a primitive type. It is more efficient in general, the API is the same with Belt_Set except its value type is fixed, and identity is not needed(using the built-in one)

Belt_SortArray A module for Array sort relevant utiliites

Belt_SortArrayInt This is a specialized module for Belt_SortArray, the docs in that module also applies here, except the comparator is fixed and inlined

Belt_SortArrayString This is a specialized module for Belt_SortArray, the docs in that module also applies here, except the comparator is fixed and inlined

Belt_internalAVLset

Belt_internalAVLtree

Belt_internalBuckets

Belt_internalBucketsType

Belt_internalMapInt

Belt_internalMapString

Belt_internalSetBuckets

Belt_internalSetInt

Belt_internalSetString

Dom

Dom_storage

Dom_storage2

Js_OO

Js_array JavaScript Array API

Js_array2 JavaScript Array API

Js_bigint JavaScript BigInt API

Js_cast Safe cast between OCaml values which share the same runtime representation

Js_console

Js_date JavaScript Date API

Js_dict

Js_exn

Js_float Provides functions for inspecting and manipulating floats

Js_global Contains functions available in the global scope (window in a browser context)

Js_int Provides functions for inspecting and manipulating ints

Js_json Efficient JSON encoding using JavaScript API

Js_list

Js_map ES6 Map API

Js_mapperRt

Js_math JavaScript Math API

Js_null Provides functionality for dealing with the 'a Js.null type

Js_null_undefined Contains functionality for dealing with values that can be both null and undefined

Js_obj Provides functions for inspecting and maniplating native JavaScript objects

Js_option

Js_promise Specialized bindings to Promise. Note: For simplicity, this binding does not track the error type, it treat it as an opaque type

Js_re Provides bindings for JavaScript Regular Expressions

Js_result

Js_set ES6 Set API

Js_string JavaScript String API

Js_string2 JavaScript String API

Js_typed_array JavaScript Typed Array API

Js_typed_array2 JavaScript Typed Array API

Js_types

Js_undefined Provides functionality for dealing with the 'a Js.undefined type

Js_vector

Js_weakmap ES6 WeakMap API

Js_weakset ES6 WeakSet API

Node Placeholder for Node bindings

Node__no_aliases Placeholder for Node bindings

Node_buffer

Node_child_process

Node_fs

Node_module

Node_path Node Path API

Node_process

Library melange.js_parser

The entry point of this library is the module: Js_parser.

Library melange.ppx

The entry point of this library is the module: Melange_ppx.

Library melange.ppxlib

The entry point of this library is the module: Melange_ppx_lib.

Library melange.ppxlib-ast

The entry point of this library is the module: Melange_ppxlib_ast.

Library melange.runtime

This library exposes the following toplevel modules:

Bs_stdlib_mini Since others depend on this file, its public mli files **should not export types** introduced here, otherwise it would cause conflicts here.

Caml

Caml_array

Caml_array_extern

Caml_bytes

Caml_exceptions

Caml_external_polyfill

Caml_float

Caml_float_extern

Caml_format

Caml_gc

Caml_hash

Caml_hash_primitive

Caml_int32

Caml_int32_extern

Caml_int64

Caml_int64_extern

Caml_io

Caml_js_exceptions

Caml_lexer

Caml_md5

Caml_module

Caml_nativeint_extern

Caml_obj

Caml_oo

Caml_oo_curry

Caml_option

Caml_parser

Caml_splice_call

Caml_string

Caml_string_extern

Caml_sys

Caml_undefined_extern

Curry

Js This library provides bindings and necessary support for JS FFI. It contains all bindings into Js namespace.